refactor(core): one shared mining module for the rules and skills miners
skills.rs carried ~150 byte-identical lines copied from rules.rs
(STOPWORDS, terms, jaccard, slugify, hash8, clustering, representative
selection) — its own header admitted the copy and undercounted it. Any
stopword/similarity/slug tweak silently applied to one miner and not
the other. The helpers now live once in crate::mining, generic over the
observation type; the per-miner bits (slug fallback word, already-
captured haystack, rules' guard inference, skills' domain union) stay
local.

Findings: D5, S5 (duplication + structure audits)
