+++ stella-core/src/rules.rs
@@ -577,92 +577,6 @@ pub struct RawObservation {
     pub memory_kind: Option<String>,
 }
 
-const STOPWORDS: &[&str] = &[
-    "the", "a", "an", "and", "or", "to", "of", "in", "on", "for", "with", "is", "are", "be",
-    "this", "that", "it", "as", "at", "by", "from", "into", "we", "you", "i", "was", "were", "has",
-    "have", "had", "not", "but", "if", "so", "then", "than", "when", "where", "which", "will",
-    "would", "should", "did",
-];
-
-/// Split text into lowercased, de-stopped terms for lexical clustering
-/// (TS: `terms`).
-fn terms(text: &str) -> Vec<String> {
-    let stopwords: HashSet<&str> = STOPWORDS.iter().copied().collect();
-    let mut out = Vec::new();
-    let mut current = String::new();
-    for ch in text.to_lowercase().chars() {
-        if ch.is_ascii_alphanumeric() || ch == '_' {
-            current.push(ch);
-        } else if !current.is_empty() {
-            if current.len() > 2 && !stopwords.contains(current.as_str()) {
-                out.push(std::mem::take(&mut current));
-            } else {
-                current.clear();
-            }
-        }
-    }
-    if current.len() > 2 && !stopwords.contains(current.as_str()) {
-        out.push(current);
-    }
-    out
-}
-
-/// Jaccard similarity of two term sets — 0 when either is empty (TS:
-/// `jaccard`).
-fn jaccard(a: &HashSet<String>, b: &HashSet<String>) -> f64 {
-    if a.is_empty() || b.is_empty() {
-        return 0.0;
-    }
-    let intersection = a.intersection(b).count();
-    let union = a.len() + b.len() - intersection;
-    if union == 0 {
-        0.0
-    } else {
-        intersection as f64 / union as f64
-    }
-}
-
-/// Filesystem/rule-id-safe slug: lowercase, alnum + dashes, capped so ids
-/// stay short (TS: `slugify`).
-fn slugify(text: &str) -> String {
-    let mut collapsed = String::new();
-    let mut prev_dash = false;
-    for ch in text.to_lowercase().chars() {
-        if ch.is_ascii_alphanumeric() {
-            collapsed.push(ch);
-            prev_dash = false;
-        } else if !prev_dash {
-            collapsed.push('-');
-            prev_dash = true;
-        }
-    }
-    let trimmed = collapsed.trim_matches('-');
-    let truncated: String = trimmed.chars().take(40).collect();
-    let truncated = truncated.trim_end_matches('-');
-    if truncated.is_empty() {
-        "lesson".to_string()
-    } else {
-        truncated.to_string()
-    }
-}
-
-/// Short deterministic content hash (FNV-1a 64-bit, lower 32 bits as 8 hex
-/// chars) so re-mining the same data yields the same candidate id (TS:
-/// `hash8`, backed by `sha1`). Not cryptographic — this is purely a stable
-/// id, not an integrity check, and no hash crate is a workspace dependency
-/// yet; FNV-1a is dependency-free and deterministic, which is all this
-/// needs.
-fn hash8(text: &str) -> String {
-    const FNV_OFFSET: u64 = 0xcbf2_9ce4_8422_2325;
-    const FNV_PRIME: u64 = 0x0000_0100_0000_01b3;
-    let mut hash = FNV_OFFSET;
-    for byte in text.as_bytes() {
-        hash ^= u64::from(*byte);
-        hash = hash.wrapping_mul(FNV_PRIME);
-    }
-    format!("{:08x}", (hash & 0xffff_ffff) as u32)
-}
-
 /// The longest common leading directory across every path, or `None` if
 /// they share none, or any path is a bare filename that can't anchor a
 /// safe glob (TS: `commonDirPrefix`).
@@ -716,66 +630,6 @@ fn infer_guard(cluster: &[RawObservation]) -> Option<RuleGuard> {
     })
 }
 
-/// The cluster's most representative wording: the most-repeated exact
-/// text, longest wins ties. `None` only for an empty cluster, which
-/// [`mine_candidates`] never constructs (TS: `representativeText`).
-fn representative_text(cluster: &[RawObservation]) -> Option<String> {
-    let mut counts: HashMap<&str, usize> = HashMap::new();
-    for o in cluster {
-        *counts.entry(o.text.as_str()).or_insert(0) += 1;
-    }
-    let mut best = cluster.first()?.text.as_str();
-    let mut best_count = 0usize;
-    for (text, count) in &counts {
-        // Most-frequent, then longest, then lexically smallest. The final
-        // lexical tiebreak is load-bearing: without it, two equal-count,
-        // equal-length texts resolve in HashMap iteration order (randomized
-        // per process), so re-mining the same observations could pick a
-        // different representative and mint a duplicate `<slug>-<hash>` file.
-        let better = *count > best_count
-            || (*count == best_count && text.len() > best.len())
-            || (*count == best_count && text.len() == best.len() && *text < best);
-        if better {
-            best = text;
-            best_count = *count;
-        }
-    }
-    Some(best.to_string())
-}
-
-/// `true` when an existing rule already says essentially the same thing
-/// (TS: `alreadyCaptured`).
-fn already_captured(text: &str, existing_rules: &[Rule], min_similarity: f64) -> bool {
-    let t: HashSet<String> = terms(text).into_iter().collect();
-    existing_rules.iter().any(|r| {
-        let rt: HashSet<String> = terms(&r.text).into_iter().collect();
-        jaccard(&t, &rt) >= min_similarity
-    })
-}
-
-/// Greedy single-pass clustering: each observation joins the first cluster
-/// whose *first* observation's term set overlaps enough with it, else
-/// starts a new one. `O(n × clusters)` — fine at CLI-local data volumes
-/// (TS: the clustering loop inside `mineCandidates`).
-fn cluster_observations(
-    observations: Vec<RawObservation>,
-    min_similarity: f64,
-) -> Vec<Vec<RawObservation>> {
-    let mut clusters: Vec<Vec<RawObservation>> = Vec::new();
-    for obs in observations {
-        let obs_terms: HashSet<String> = terms(&obs.text).into_iter().collect();
-        let home = clusters.iter().position(|c| {
-            let head_terms: HashSet<String> = terms(&c[0].text).into_iter().collect();
-            jaccard(&obs_terms, &head_terms) >= min_similarity
-        });
-        match home {
-            Some(idx) => clusters[idx].push(obs),
-            None => clusters.push(vec![obs]),
-        }
-    }
-    clusters
-}
-
 /// Mine [`RawObservation`]s into ranked rule-promotion candidates. A
 /// cluster of similar-enough observations (Jaccard ≥
 /// `config.min_similarity`) qualifies when either it recurred at least
@@ -787,7 +641,8 @@ pub fn mine_candidates(
     existing_rules: &[Rule],
     config: &MineConfig,
 ) -> Vec<RuleCandidate> {
-    let clusters = cluster_observations(observations, config.min_similarity);
+    let clusters =
+        crate::mining::cluster_observations(observations, config.min_similarity, |o| &o.text);
     let mut candidates: Vec<RuleCandidate> = Vec::new();
 
     for cluster in clusters {
@@ -795,10 +650,14 @@ pub fn mine_candidates(
         if cluster.len() < config.min_occurrences && !salient {
             continue;
         }
-        let Some(text) = representative_text(&cluster) else {
+        let Some(text) = crate::mining::representative_text(&cluster, |o| &o.text) else {
             continue;
         };
-        if already_captured(&text, existing_rules, config.min_similarity) {
+        if crate::mining::already_captured(
+            &text,
+            existing_rules.iter().map(|r| r.text.clone()),
+            config.min_similarity,
+        ) {
             continue;
         }
 
@@ -824,7 +683,11 @@ pub fn mine_candidates(
         };
 
         candidates.push(RuleCandidate {
-            id: format!("{}-{}", slugify(&text), hash8(&text)),
+            id: format!(
+                "{}-{}",
+                crate::mining::slugify(&text, "lesson"),
+                crate::mining::hash8(&text)
+            ),
             description: format!(
                 "Promoted from {occurrences} recurring observation{plural}{salience_note}."
             ),
